Add -Wconversion to extra-warnings

+++ tools/thumbnail.c
@@ -47,7 +47,7 @@
 
 #ifndef TIFFhowmany8
 #define TIFFhowmany8(x)                                                        \
-    (((x)&0x07) ? ((uint32_t)(x) >> 3) + 1 : (uint32_t)(x) >> 3)
+    ((int)((((x)&0x07) ? ((uint32_t)(x) >> 3) + 1 : (uint32_t)(x) >> 3)))
 #endif
 
 typedef enum
@@ -87,10 +87,10 @@ int main(int argc, char *argv[])
         switch (c)
         {
             case 'w':
-                tnw = strtoul(optarg, NULL, 0);
+                tnw = (uint32_t)strtoul(optarg, NULL, 0);
                 break;
             case 'h':
-                tnh = strtoul(optarg, NULL, 0);
+                tnh = (uint32_t)strtoul(optarg, NULL, 0);
                 break;
             case 'c':
                 contrast = streq(optarg, "exp50")    ? EXP50
@@ -475,7 +475,7 @@ static void setupBitsTables(void)
 
 static int clamp(float v, int low, int high)
 {
-    return (v < low ? low : v > high ? high : (int)v);
+    return (v < (float)low ? low : v > (float)high ? high : (int)v);
 }
 
 #ifndef M_E
@@ -558,8 +558,8 @@ static void setupStepTables(uint32_t sw)
 {
     if (stepSrcWidth != sw || stepDstWidth != tnw)
     {
-        int step = sw;
-        int limit = tnw;
+        int step = (int)sw;
+        int limit = (int)tnw;
         int err = 0;
         uint32_t sx = 0;
         uint32_t x;
@@ -575,15 +575,15 @@ static void setupStepTables(uint32_t sw)
                 sx++;
             }
             rowoff[x] = sx0 >> 3;
-            fw = sx - sx0; /* width */
+            fw = (int)(sx - sx0); /* width */
             b = (fw < 8) ? (uint8_t)(0xff << (8 - fw)) : (uint8_t)0xff;
             src0[x] = b >> (sx0 & 7);
-            fw -= 8 - (sx0 & 7);
+            fw -= (int)(8 - (sx0 & 7));
             if (fw < 0)
                 fw = 0;
             src1[x] = (uint8_t)(fw >> 3);
             fw -= (fw >> 3) << 3;
-            src2[x] = 0xff << (8 - fw);
+            src2[x] = (uint8_t)(0xff << (8 - fw));
         }
         stepSrcWidth = sw;
         stepDstWidth = tnw;
@@ -653,8 +653,8 @@ static void setrow(uint8_t *row, uint32_t nrows, const uint8_t *rows[])
  */
 static void setImage1(const uint8_t *br, uint32_t rw, uint32_t rh)
 {
-    int step = rh;
-    int limit = tnh;
+    int step = (int)rh;
+    int limit = (int)tnh;
     int err = 0;
     int bpr = TIFFhowmany8(rw);
     int sy = 0;
